Convert sessions integration tests
- reengineer these tests to avoid the weird
  try/rescue block
